refactor(engine): split mission-finalize into one policy per module
Move-only: the facade keeps both public exports and the stop-reason
routing table (arm order is the contract); each terminal/park policy owns
its own module under mission-finalize/. Behavior, log events and return
values unchanged; full root suite green on this tree (16586 tests). Em
dashes in moved comments became hyphens per the writing decree.
